Added players age days (big thx to Flattermann again). Original comment: This patch implements support for the player's AgeDays.
AgeDays are shown in:
- Playeroverview
- Lineup
- Transferscout
- Transfer-Miniscout

The age is now shown in the format "years.days".

The time since your last HRF update is taken into account, so that the age
of your players should be correct, even if the last HRF update was some
days ago.
If the player had his birthday since the last HRF update, its age is shown
as "years.days(+1)".

We need to do a DB update (v6) to support the new AgeDays column. (Tables
SPIELER & SCOUT)

git-svn-id: https://ho1.svn.sourceforge.net/svnroot/ho1@73 b8de521a-fa13-0410-8113-fc7ffabbd59a