A commit waiting for oauth to be done first.
Training defaults are new, and the way this option menu works is modified.
Prediction Offset is fixed for 553, which gave a few changes in lineup and spielerposition among others.
A temp player bug in lineup gui was also fixed in through a mod in SpielerPosition.
Parsing of HT data is updated to correct buggy matchlineup api regarding individual orders.


git-svn-id: file:///home/coren/project/backup_ho/trunk/Core@857 b8de521a-fa13-0410-8113-fc7ffabbd59a