Add sources to build Shell32 JNI wrapper
Java application has to call native function to get application data
path in Windows. There exists JNA helper library to do the call but that
would be quite big dependency to add for a single native call. That
makes it desirable to use JNI interface directly. That requires small
amount of C code to forward java call to the native function.