Don't let login wait dialog stay visible after use
Swing may only be accessed from event thread. But we know that event
loop is currently blocked that makes it enough to synchronize access in
our extended class to underlying race implementations.

This fixes issue that WaitDialog randomly staying in screen that was
caused by race between rendering and closing the window.

git-svn-id: https://ho1.svn.sourceforge.net/svnroot/ho1/trunk@2040 b8de521a-fa13-0410-8113-fc7ffabbd59a